drone that rescale !

+++ src/models.py
@@ -1,6 +1,6 @@
 class NodeData:
     def __init__(self, name: str, x: int, y: int, zone: str,
-                 color: str | None, max_drones: float):
+                 color: str | None, max_drones: int):
         self.name = name
         self.x = x
         self.y = y
